chore: rename file oudscoloredbox to lightdarkbox
