+++ app/lib/ui/components/tag/input_tag_demo_screen.dart
@@ -23,9 +23,9 @@ import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_section
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_switch.dart';
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_textfield.dart';
 import 'package:ouds_flutter_demo/ui/utilities/detail_screen_header.dart';
+import 'package:ouds_flutter_demo/ui/utilities/light_dark_box.dart';
 import 'package:ouds_flutter_demo/ui/utilities/reference_design_version_component.dart';
 import 'package:ouds_flutter_demo/ui/utilities/sheets_bottom/ouds_sheets_bottom.dart';
-import 'package:ouds_flutter_demo/ui/utilities/theme_colored_box.dart';
 import 'package:ouds_theme_contract/ouds_component_version.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
 import 'package:provider/provider.dart';
@@ -119,11 +119,11 @@ class _InputTagDemoState extends State<_InputTagDemo> {
     themeController = Provider.of<ThemeController>(context, listen: true);
 
     return LightDarkBox(
-          child: OudsInputTag(
-            label: customizationState?.labelText ?? "",
-            onPressed: customizationState?.hasEnabled == true ? () {} : null,
-          ),
-        );
+      child: OudsInputTag(
+        label: customizationState?.labelText ?? "",
+        onPressed: customizationState?.hasEnabled == true ? () {} : null,
+      ),
+    );
   }
 }
